NOTE: If you see errors after this change, it is most likely a simple issue solved by refreshing nodes because I renamed an actor.
Made sure the children of the old base projectile weapon now work with the new one. Fixed the roulette weapon errors (it was trying to use old variables that no longer existed), and also added the new ones inside it so it gets the stats correctly. Renamed "BP_BaseProjectileWeapon" to "BP_BaseWeapon". Some other minor changes.